do Jay's suggestion with posix/os API naming & layout
