+++ std/os/test.zig
@@ -1,5 +1,6 @@
 const std = @import("../std.zig");
 const os = std.os;
+const testing = std.testing;
 const expect = std.testing.expect;
 const io = std.io;
 const mem = std.mem;
@@ -126,4 +127,25 @@ fn testTls(context: void) void {
     if (x != 1234) @panic("bad start value");
     x += 1;
     if (x != 1235) @panic("bad end value");
+}
+
+test "getrandom" {
+    var buf_a: [50]u8 = undefined;
+    var buf_b: [50]u8 = undefined;
+    try os.getrandom(&buf_a);
+    try os.getrandom(&buf_b);
+    // If this test fails the chance is significantly higher that there is a bug than
+    // that two sets of 50 bytes were equal.
+    expect(!mem.eql(u8, buf_a, buf_b));
+}
+
+test "getcwd" {
+    // at least call it so it gets compiled
+    var buf: [std.fs.MAX_PATH_BYTES]u8 = undefined;
+    _ = os.getcwd(&buf) catch {};
+}
+
+test "realpath" {
+    var buf: [std.fs.MAX_PATH_BYTES]u8 = undefined;
+    testing.expectError(error.FileNotFound, os.realpath("definitely_bogus_does_not_exist1234", &buf));
 }
